networkstate reputation i32 (#2430)

+++ packages/types/src/interfaces/system/definitions.ts
@@ -222,7 +222,7 @@ export default {
     },
     NetworkStatePeersetInfo: {
       connected: 'bool',
-      reputation: 'u64'
+      reputation: 'i32'
     },
     NodeRole: {
       _enum: {
